updated kill enhanced script and changed folder path
I added some more features, such as more data on the player killed marker, distance, and weapon used to kill the player is now on their death marker. I also add ace logic to check if the player is unconscious. If so, a yellow warning marker is placed and a message is broadcast. If they die or regain consciousness, the marker will be removed. I added CBA settings for everything added, to make it modular. Turn it on, turn it off, whatever floats your boat...